Remove an old gettext initializer that's unused
Other providers don't have this so I think we can remove it.

Note, this is not compatible with rails 7.  We'd have to either require
the file before using the constant or possibly do this in an after_initialize
block or somehow tell it to autoload this once.  Rails 7 doesn't want to autoload
in initializers because they're not reloaded when the application code is reloaded.